# cnj-openapi-backend-micro/src/main/resources/META-INF/openapi.yaml
openapi: 3.0.0
info:
  title: cnj-open-backend-micro
  description: Cloud native Java backend based on Eclipse Microprofile using MP OpenAPI
    feature to document REST APIs
  contact:
    url: https://github.com/msgoat/cnj-openapi
  version: 4.0.0
servers:
  - url: https://apps.cloudtrain.aws.msgoat.eu/cloudtrain-int/cnj-openapi-backend-micro
paths:
  /api/v1/tasks:
    get:
      summary: returns all tasks
      operationId: getAllTasks
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schema/Task'
          description: body contains all available tasks
      security:
        - oidc:
            - openid
            - microprofile-jwt
    post:
      summary: creates a new task based on the given task data and returns its location
        URI
      operationId: addTask
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Task'
      responses:
        "201":
          headers:
            Location:
              description: URI of newly created task
              style: simple
          description: task could be created successfully; header location contains
            URI of newly created task; body is empty
        "400":
          description: failed to create new task due to invalid task data
        "500":
          description: failed to create new task due to internal error
      security:
        - oidc:
            - openid
            - microprofile-jwt
  /api/v1/tasks/{taskId}:
    get:
      summary: returns the task with the given task ID
      operationId: getTask
      parameters:
        - name: taskId
          in: path
          required: true
          style: simple
          schema:
            type: object
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schema/Task'
          description: body contains the task with the given task ID
        "400":
          description: unable to find a task with the given task ID; body is empty
      security:
        - oidc:
            - openid
            - microprofile-jwt
    put:
      summary: updates the given task
      operationId: modifyTask
      parameters:
        - name: taskId
          in: path
          required: true
          style: simple
          schema:
            type: object
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Task'
      responses:
        "204":
          description: task could be updated successfully; body is empty
        "400":
          description: failed to update task due to invalid task data
        "500":
          description: failed to create new task due to internal error
      security:
        - oidc:
            - openid
            - microprofile-jwt
    delete:
      summary: deletes the given task
      operationId: removeTask
      parameters:
        - name: taskId
          in: path
          required: true
          style: simple
          schema:
            type: object
      responses:
        "204":
          description: task could be deleted successfully; body is empty
        "500":
          description: failed to delete task due to internal error
      security:
        - oidc:
            - openid
            - microprofile-jwt
components:
  schemas:
    Task:
      properties:
        affectedApplicationId:
          maxLength: 32
          type: string
          description: application-ID of the application this task is related to
        lastModifiedAt:
          type: string
          description: date/time when this task has been modified
          format: date-time
          readOnly: true
        subject:
          maxLength: 80
          type: string
          description: single-line description or summary of what this task is about
        dueDate:
          type: string
          description: date/time when this task is supposed to be completed
          format: date-time
        lastModifiedBy:
          maxLength: 31
          type: string
          description: user-ID of the last user who modified this task
          readOnly: true
        description:
          maxLength: 1024
          type: string
          description: detailed description of this task
        completionRate:
          maximum: 100
          minimum: 0
          type: integer
          description: completion rate in percent, ranges from 0 to 100
        priority:
          type: object
          description: priority
          enum:
            - UNDEFINED
            - LOW
            - MEDIUM
            - HIGH
            - CRITICAL
        completedByUserId:
          maxLength: 16
          type: string
          description: user-ID of participant who completed this task
        createdAt:
          type: string
          description: date/time when this task has been created
          format: date-time
          readOnly: true
        submitterUserId:
          type: string
          description: user-ID of participant who submitted this task
        affectedResource:
          maxLength: 256
          type: string
          description: application resource that this task is referring to
        createdBy:
          maxLength: 31
          type: string
          description: user-ID of the user who created this task
          readOnly: true
        actualEffort:
          type: integer
          description: actual effort in hours to complete this task
        affectedProjectId:
          maxLength: 32
          type: string
          description: project-ID of the project this task is related to
        completionDate:
          type: object
          description: date/time when this task has been completed
        id:
          type: string
          description: unique identifier of this task
          format: uuid
          readOnly: true
        category:
          type: object
          description: 'groups task into specific categories (see: TaskCategory)'
          enum:
            - UNDEFINED
            - BUGFIX
            - REFACTORING
            - NEW_FEATURE
            - PERFORMANCE_IMPROVEMENT
            - RELEASE_MANAGEMENT
            - QUALITY_ASSURANCE
            - BUILD_FAILURE
            - COMMUNICATION
        submittedAt:
          type: string
          description: date/time when this task has been submitted
          format: date-time
        responsibleUserId:
          maxLength: 16
          type: string
          description: user-ID of participant who is currently responsible for the
            completion of this task
        estimatedEffort:
          type: integer
          description: estimated effort in hours to complete this task
        lifeCycleState:
          type: object
          description: status of this task
          enum:
            - UNDEFINED
            - OPEN_UNDER_WORK
            - OPEN_RUNNING
            - CLOSED_COMPLETED
            - CLOSED_REVOKED
            - CLOSED_CANCELLED
        affectedModule:
          maxLength: 32
          type: string
          description: name of the logical module this task is related to
      description: represents a task
    TaskCategory:
      description: enumeration representing the various categories of a task
      enum:
        - UNDEFINED
        - BUGFIX
        - REFACTORING
        - NEW_FEATURE
        - PERFORMANCE_IMPROVEMENT
        - RELEASE_MANAGEMENT
        - QUALITY_ASSURANCE
        - BUILD_FAILURE
        - COMMUNICATION
    TaskLifeCycleState:
      description: enumeration representing the life-cycle of a task
      enum:
        - UNDEFINED
        - OPEN_UNDER_WORK
        - OPEN_RUNNING
        - CLOSED_COMPLETED
        - CLOSED_REVOKED
        - CLOSED_CANCELLED
    TaskPriority:
      description: enumeration representing the priority of a task
      enum:
        - UNDEFINED
        - LOW
        - MEDIUM
        - HIGH
        - CRITICAL
  securitySchemes:
    oidc:
      type: openIdConnect
      openIdConnectUrl: https://oidc.cloudtrain.aws.msgoat.eu/auth/realms/cloudtrain/.well-known/openid-configuration